[WIP] refactor to smaller components for readability
